M117/Issue #181 - Added ability to set windows for referenced coverages.
