+++ coverage_model/db_backed_metadata.py
@@ -9,13 +9,14 @@
 
 from ooi.logging import log
 import os
-from coverage_model.metadata import MetadataManager
-from coverage_model.utils import hash_any
-from coverage_model.persistence_helpers import RTreeProxy, pack, unpack
 from coverage_model.basic_types import Dictable
-from coverage_model.db_connectors import DBFactory
+from coverage_model.coverages.coverage_extents import ReferenceCoverageExtents
 from coverage_model.data_span import SpanStatsCollection, SpanCollectionByFile, SpanStats
-from coverage_model.persistence_helpers import MasterManager, BaseManager
+from coverage_model.db_connectors import DBFactory
+from coverage_model.metadata import MetadataManager
+from coverage_model.persistence_helpers import RTreeProxy, pack, unpack, MasterManager, BaseManager
+from coverage_model.util.jsonable import Jsonable
+from coverage_model.utils import hash_any
 
 
 class DbBackedMetadataManager(MetadataManager):
